Fixes for last changes

+++ ios/Doglyad/Domain/Config/UltrasoundConfig.swift
@@ -1,6 +1,6 @@
 struct UltrasoundConfig: Codable {
     let scanPhotoMaxNumber: Int
-    let scanPhotoResizeMaxDimension: Int
+    let scanPhotoResizeMaxDimension: Double
     let scanPhotoCompressionQuality: Double
     let defaultPatientDateOfBirthGap: Int
     let defaultPatientHeightCM: Int
